python: move ipc code to own module
